Mention `write` from the `format` command
When trying to figure out how to format a new Ruby project, we were choosing between `@prettier/plugin-ruby`, this project, and Rufo. 

At first, we thought this project didn't have built-in support for writing to disk at all, so we almost switched to Rufo (or prettier-ruby, despite that library's admonition to use this gem). Only when I said, "no, Kevin wouldn't ship a library like this without a `--write` facility!" did we keep scrolling and find the `write` command.